Criado a camada de recurso do produto
